transformed delete in component, added to delete animal in profile page
