+++ client/src/components/profile/profile_header/ProfileHeader.jsx
@@ -1,7 +1,9 @@
+import { useNavigate, useParams } from "react-router-dom";
 import PropTypes from "prop-types";
 import styles from "./ProfileHeader.module.css";
 import DeleteProfile from "./delete_profile/DeleteProfile";
 import InputFile from "../input_file/InputFile";
+import notify from "../../../utils/notify";
 
 function ProfileHeader({
   username,
@@ -14,6 +16,37 @@ function ProfileHeader({
   changeAvatar = null,
 }) {
   const { avatar } = customer;
+  const navigate = useNavigate();
+  const { id } = useParams();
+  const URL = import.meta.env.VITE_API_URL;
+
+  // Delete profile
+  const deleteprofile = async () => {
+    try {
+      const response = await fetch(`${URL}user/${id}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({id}),
+      });
+
+      if (response.status === 204) {
+        notify("Le profile à été supprimé", "success");
+        return navigate("/");
+      }
+      throw new Error("Registration error");
+    } catch (err) {
+      console.error("Fetch error:", err);
+      notify(
+        "Erreur lors de la suppression du profil. Veuillez réessayer plus tard.",
+        "error"
+      );
+      return {
+        error: "An error occurred during deletion. Please try again later.",
+      };
+    }
+  };
 
   return (
     <header className={styles.profilePageHeader}>
@@ -51,7 +84,12 @@ function ProfileHeader({
             >
               {isEditMode === true ? "Sauvegarder" : "Modifier"}
             </button>
-            {isEditMode === false && <DeleteProfile />}
+            {isEditMode === false && (
+              <DeleteProfile
+                 text="Êtes vous sur de vouloir supprimer votre compte ?"
+                 deleteOnClick={deleteprofile}
+               />
+            )}
           </div>
         )}
       </div>
